webapp/webapp/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.1
python-multipart==0.0.6
google-generativeai==0.3.2
aiofiles==23.2.0